backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dateutil==2.8.2
aiofiles==23.2.1
pydantic-settings==2.1.0
alembic==1.13.1
email-validator==2.1.0
